Fix prototype issue with Android Browser
Generating a new Ractive object fails on Android browser with an error message saying "render" is not defined. 
I found that none of the prototype functions actually got added to the object on Android.

Traced the problem to the way the prototype was added to the Ractive object. It used defineProperties with a line where the property name and the variable name was both "prototype". This threw the Android browser off.
I changed the variable name and did a direct value assignment. The changed code worked in Firefox 28, Chrome 31, Android 4.0 and Android 2.3.