check for Promise on window
We can’t just check for `typeof Promise === “function”` for reliance on
native promises, because `Promise` is declared and therefore will be
considered undefined, and at best it will be self-referential. If we
check against `self.Promise`, we’ll know that we’re checking in at
least a worker or window scope and be able to utilize those promises.