Trim down root package.json down to building only
Following the lead of Lodash. Since the package.json that gets published
isn't the one on the root, this commit removes the unneeded portions of
the file. Also makes the package.json private to avoid it from being
published from there.

Should anything need be added from a publishing perspective, add it to the
manifests over at scripts/templates.