Computed properties are unwrapped by ractive.get() in children components
The incorrect behavior described in issue #3137 still existed when a
child of a computed property was passed to a child component.

An instance of LinkModel where created. The LinkModel called
the `get()` method of ComputationChild with the parameter `shouldCapture`
set to `false` and parameter `opts` set to `{unwrap: true}`.

This resulted in the value not being unwrapped because of a logic error
I introduced in `ComputationChild.get()` in the pull-request #3138.

This commit corrects the logic error and adds a unit test to be sure
ractive.get() has the same behavior regardless of whether it is
called from a root component or from a child component.